[skip ci] Gate: Update docs w/ safer local command
Previously if you used the documented code validation command,
you would end up with your repository SELinux labels being changed.  In
many cases this is not desirable.  Update the documentation such that
the original source of the repository isn't touched in any way (prior
to making a copy of it in the container).

Also add a small (but important) note about building the container
image.

Signed-off-by: Chris Evich <cevich@redhat.com>