create: don't append version to AppArmor profile
Do not append `-$VERSION` to Podman's default AppArmor profile.  The
initial intention was to immediately load new AppArmor profiles when
Podman gets updated.  However, this breaks when the host is rebooted
and pre-existing containers cause runc to load the profile of the
previous version of Podman which is not loaded anymore.

Note that this means that we need to make users aware whenever the
default AppArmor profile is changed.

Fixes: #2107
Signed-off-by: Valentin Rothberg <rothberg@redhat.com>