Fix error code retrieval for podman start --attach
When we start a container with 'podman run' and it exits and is
removed before we get the code, we grab its exit code from the
Conmon exit file. Podman start --attach wants to do the same, but
was missing the logic. Fix that here.

Also, remove some --rm handling leftover in start. Don't need it
anymore, we have the cleanup process now.

Signed-off-by: Matthew Heon <mheon@redhat.com>