Recursively stop a pod.
Before, there existed an edge case where a container in a pod stop would fail to stop, which would allow it to stop after the pod infra container. This breaks the invariant that the infra container should be running while any container is running. fix this by recursively stopping in a pod.

Signed-off-by: Peter Hunt <pehunt@redhat.com>