Cirrus: Notify on some post-merge failures
There is currently no simple way in Cirrus to perform actions when a
build or task has failed.  However, this means it's easily possible for
a flaking test or conflict to become part of master.  Add a
wrapper-script to monitor for some of these conditions, and post a
notice on IRC if/when it happens.

***Note:*** This is temporary.  An upstream RFE, marked "Comming Soon"
will provide a complete solution when it's released.
https://github.com/cirruslabs/cirrus-ci-docs/issues/55

Signed-off-by: Chris Evich <cevich@redhat.com>